# src/openapi.yml
openapi: 3.0.3
info:
  title: CarX - Car Rental API
  version: 0.0.1
  description: |
    **CarX** is a car-sharing API designed to function like Turo, offering a seamless platform for users to find, book, and manage car rentals. With CarX, users can browse a wide range of available vehicles, select their desired car, and book it for their next trip. Once booked, each reservation is tracked as a trip within the system, allowing users to manage their rentals efficiently.

    The API integrates with Stripe for secure and reliable payment processing, ensuring a smooth checkout experience for all transactions. Whether you're looking for a short-term rental or planning a road trip, CarX provides a user-friendly and flexible solution for car sharing
    
    CarX makes car rentals easy, reliable, and accessible, providing a modern alternative to traditional car rental services.
  contact:
    name: Hassan Attar
    email: h.a.develops@gmail.com
    url: https://hassan-attar.com
  license:
    name: CarX License
    url: https://opensource.org/license/mit
paths:
  /cars:
    get:
      summary: Retrieve a list of cars
      operationId: getCars
      tags:
        - Cars
      parameters:
        - name: availableFrom
          in: query
          required: false
          description: |
            The starting date in ISO 8601 date format. 
            - Must be a valid ISO 8601 date string.
            - Defaults to the current date and time (`now`) if left blank.
            - Cannot be earlier than the current date and time.
          schema:
            type: string
            format: date-time
            description: The starting date in ISO 8601 date format. (defaults to now)
          examples:
            LeaveBlank:
              value: ""
              summary: Leave it blank (defaults to now on the server) or write custom value
            SpecificDate:
              value: '2024-10-10T00:00:00.000Z'
              summary: a valid ISO 8601 date (modify it as you wish)
        - name: availableTo
          in: query
          required: false
          description: |
            The ending date in ISO 8601 date format.
            - Must be a valid ISO 8601 date string.
            - Defaults to three days after `availableFrom` if `availableFrom` is provided or 3 days after now.
            - Must be greater than the `availableFrom` date.
          schema:
            type: string
            format: date-time
            description: The ending date in ISO 8601 date format. (defaults to now + 3 days)
          examples:
            LeaveBlank:
              value: ""
              summary: Leave it blank (defaults to 3 days after now on the server) or write custom value
            SpecificDate:
              value: '2024-10-13T00:00:00.000Z'
              summary: a valid ISO 8601 date (modify it as you wish)
        - name: country
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
            description: |
              The country code or name. Must be a valid ISO 3166-1 alpha-2, alpha-3, or numeric code.
              The server will validate the provided value.
              If you provide a valid country name, (e.g. United States for USA), server will try to cast it to ISO 3166-1.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            validAlpha2:
              summary: ISO 3166-1 Alpha-2 Code
              value: "US"
            validAlpha3:
              summary: ISO 3166-1 Alpha-3 Code (Canada)
              value: "CAN"
            validNumericCode:
              summary: ISO 3166-1 numeric Code (USA)
              value: 840
            validName:
              summary: A valid name for USA
              value: "United States"
            anotherValidName:
              summary: Another valid name for USA
              value: "United States of America"
            invalid:
              summary: Invalid Country
              value: "XYZ"
        - name: region
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
            description: |
              The region or subdivision code. Must be a valid ISO 3166-2 code for subdivisions. 
              Requires a valid country parameter to be specified. 
              The server will validate that the region belongs to the specified country.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            validUS:
              summary: Valid Region Code for USA as country (2 letter code)
              value: "WA"
            validUSCA:
              summary: Valid Region Code for USA as country (full name)
              value: "Washington"
            validCA:
              summary: Valid Region Code for Canada as country (2 letter code)
              value: "bc"
            validCAA:
              summary: Valid Region Code for Canada (full name)
              value: "British Columbia"
            invalid:
              summary: Invalid Region Code for Canada as country
              value: "CA"
        - name: city
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
            description: The city name.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            inUS:
              summary: City in US
              value: los angeles
            inCA:
              summary: City in Canada
              value: vancouver
        - name: lat
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
          description: >
            Latitude value. Must be between -90 and 90. 
            If specified, `lng` must also be provided.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            VancouverLat:
              value: 49.232504
              summary: Lat for Vancouver, Canada
            SeattleLat:
              value: 47.6114
              summary: Lat for Seattle, US
        - name: lng
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
          description: >
            Longitude value. Must be between -180 and 180. 
            If specified, `lat` must also be provided.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            VancouverLng:
              value: -123.138901
              summary: Lng for Vancouver, Canada
            SeattleLng:
              value: -122.3305
              summary: Lng for Seattle, US
        - name: radius
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 1
          description: >
            Search radius in kilometers. Must be at least 1. 
            This field is only allowed when both `lat` and `lng` are present.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            within50KM:
              summary: Within 50 KM proximity
              value: 50
            within500KM:
              summary: Within 500 KM proximity
              value: 500
        - name: priceMin
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
            description: The minimum price filter. Must be a positive number.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            Default:
              value: 60.99
              summary: Example Price

        - name: priceMax
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
            description: The maximum price filter. Must be a positive number.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            Default:
              value: 100.00
              summary: Example Price

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: >
            The maximum number of items per page. Must be a positive integer 
            up to 100. Defaults to 10 if not specified.
          examples:
            LeaveBlank:
              summary: Leave it blank (defaults to 10) or write custom value
              value: ""
            limit2:
              summary: Limit of 2
              value: 2
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: >
            The page number to retrieve. Must be a positive integer. 
            Defaults to 1 if not specified.
          examples:
            LeaveBlank:
              summary: Leave it blank or write custom value
              value: ""
            page2:
              value: 2
              summary: Go Page 2
      description: |
        This endpoint allows you to retrieve a list of available cars based on various filters such as availability dates, location, and price range.
        - Use `availableFrom` and `availableTo` to filter cars by their availability period.
        - Location-based filtering can be done using `country`, `region`, `city`, or precise `lat` and `lng` coordinates along with a `radius`.
        - Price range can be adjusted using `priceMin` and `priceMax`.
        - Use pagination with `limit` and `page` to control the number of results returned.
        
        **By default, you get all cars that are available from now until next 3 days.**
      responses:
        '200':
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
          links:
            GetCarDetails:
              operationId: getCarDetails
              parameters:
                carId: $response.body[*].carId
        '422':
          description: Validation Error
          $ref: "#/components/responses/ValidationError"
        '500':
          description: "Internal Server Error"
          $ref: "#/components/responses/InternalServerError"
...
servers:
  - url: https://carx.hassan-attar.com/api/{version}
    description: "Production Server"
    variables:
      version:
        default: v1
        enum:
          - v1
  - url: http://localhost:8080/api/{version}
    description: "Local Development Server"
    variables:
      version:
        default: v1
        enum:
          - v1
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: _sid
      description: |
        ## Ways to Obtain the Authentication Cookie:
        
        ### 1. **Through the Frontend Application:**
           - You can log in through our **[frontend application](https://carx.hassan-attar.com)**. Once logged in, you can return to this API docs to continue browsing
        
        ### 2. **Through API Endpoints:**
           - You can also log in or sign up directly via the API endpoints:
             - **Sign Up**: Use the `/auth/signup` endpoint to create a new account and log in.
             - **Log In**: Use the `/auth/login` endpoint to log in with your credentials.
        
        `Once Logged in, The session cookie will attach to all of your requests automatically so you can browse the API freely.`
  schemas:
    Location:
      type: object
      description: Location of the car
      properties:
        crs:
          type: object
          properties:
            type:
              type: string
            properties:
              type: object
              properties:
                name:
                  type: string
                  description: Coordinate reference system name
        type:
          type: string
          description: Location type (e.g., Point)
        coordinates:
          type: array
          items:
            type: number
          description: Longitude and latitude of the car
          minItems: 2
          maxItems: 2
          example: [ -123.053418, 49.232504 ]
    FuelType:
      type: string
      description: Type of fuel the car uses
      enum: [ "gasoline", "hybrid", "electric" ]
    CarType:
      type: string
      description: Type of vehicle
      enum: [ "sedan", "suv", "truck", "van", "coupe" ]
    TransmissionType:
      type: string
      enum:
        - auto
        - manual
      description: Type of transmission
...
    BaseCar:
      type: object
      properties:
        carId:
          $ref: "#/components/schemas/UUID"
        make:
          type: string
          description: Make of the car
        model:
          type: string
          description: Model of the car
        headerImage:
          type: string
          format: uri
          description: URL of the car's header image
        year:
          type: integer
          description: Year of manufacture
        city:
          type: string
          description: City where the car is located
        region:
          type: string
          description: Region or state of the car's location
        country:
          type: string
          description: Country code of the car's location
    Car:
      allOf:
        - $ref: "#/components/schemas/BaseCar"
        - type: object
          properties:
            transmission:
              $ref: '#/components/schemas/TransmissionType'
            type:
              $ref: '#/components/schemas/CarType'
            description:
              type: string
              description: Description of the car
            distanceIncludedKm:
              type: integer
              description: Kilometers included in the rental price
            fuelType:
              $ref: '#/components/schemas/FuelType'
            galleryImages:
              type: array
              items:
                type: string
                format: uri
              nullable: true
              description: List of URLs of gallery images
            location:
              $ref: '#/components/schemas/Location'
            avgRating:
              type: number
              nullable: true
              description: Average rating of the car
            tripCount:
              type: integer
              nullable: true
              description: Number of trips the car has been rented for
            createdAt:
              type: string
              format: date-time
              description: Timestamp of when the car was created
            updatedAt:
              type: string
              format: date-time
              description: Timestamp of the last update
...
    BaseError:
      type: object
      properties:
        error:
          type: string
          description: The type of error that occurred.
        message:
          type: string
          description: A descriptive message providing more details about the error.
        details:
          type: array
          description: Additional details about the error, if available.
          items:
            type: object
            properties:
              message:
                type: string
                description: A descriptive message providing more details.
            additionalProperties: true
          nullable: true
      required:
        - error
        - message
...
    UUID:
      description: A generic text field to regulate the expected length and format of the input fields throughout the API
      type: string
      format: uuid
      readOnly: true
      example: 1951a7e9-b2be-4797-9430-14882da9e883
  responses:
    InternalServerError:
      description: Internal Server Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseError"
          example:
            error: InternalServerError
            message: An unexpected error occurred
            details: [ ]
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseError"
          example:
            error: ValidationError
            message: Invalid query parameters; Please fix your mistakes and try again.
            details:
              - message: "You must provide a country. Please specify the country in order to specify the region."
                path: [ "region" ]
                type: "region.no-country"
                "context":
                  label: "region"
                  value: "CA"
                  key: "region"
...